+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 `1.0.15`
 
 - Added class file version test to ensure correct class files are built.
+- Added `ClassFile` interface for class files utility methods.
+- Added `ClassFile.Version` record which represents a class file version.
+- Added `maven-properties-plugin` to `pom.xml` to output all Maven properties to `project.properties`.
 
 ---